Moving emap from experimental to algorithms, verbose option in write_verilog
