+++ experiments/emap.cpp
@@ -30,12 +30,11 @@
 #include <fmt/format.h>
 #include <lorina/aiger.hpp>
 #include <lorina/genlib.hpp>
-#include <mockturtle/algorithms/experimental/emap.hpp>
+#include <mockturtle/algorithms/emap.hpp>
 #include <mockturtle/io/aiger_reader.hpp>
 #include <mockturtle/io/genlib_reader.hpp>
 #include <mockturtle/networks/aig.hpp>
 #include <mockturtle/networks/block.hpp>
-#include <mockturtle/networks/klut.hpp>
 #include <mockturtle/utils/name_utils.hpp>
 #include <mockturtle/utils/tech_library.hpp>
 #include <mockturtle/views/cell_view.hpp>
@@ -68,7 +67,7 @@ int main()
   tps.verbose = true;
   tech_library tech_lib( gates, tps );
 
-  for ( auto const& benchmark : iscas_benchmarks() )
+  for ( auto const& benchmark : epfl_benchmarks() )
   {
     fmt::print( "[i] processing {}\n", benchmark );
